backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
# samsungtvws==2.6.0
git+https://github.com/NickWaterton/samsung-tv-ws-api.git
